Split bulk_mutations for products into multiple files
