+++ saleor/graphql/product/bulk_mutations/__init__.py
@@ -1,0 +1,23 @@
+from .category_bulk_delete import CategoryBulkDelete
+from .collection_bulk_delete import CollectionBulkDelete
+from .product_bulk_delete import ProductBulkDelete
+from .product_media_bulk_delete import ProductMediaBulkDelete
+from .product_type_bulk_delete import ProductTypeBulkDelete
+from .product_variant_bulk_create import ProductVariantBulkCreate
+from .product_variant_bulk_delete import ProductVariantBulkDelete
+from .product_variant_stocks_create import ProductVariantStocksCreate
+from .product_variant_stocks_delete import ProductVariantStocksDelete
+from .product_variant_stocks_update import ProductVariantStocksUpdate
+
+__all__ = [
+    "CategoryBulkDelete",
+    "CollectionBulkDelete",
+    "ProductBulkDelete",
+    "ProductMediaBulkDelete",
+    "ProductTypeBulkDelete",
+    "ProductVariantBulkCreate",
+    "ProductVariantBulkDelete",
+    "ProductVariantStocksCreate",
+    "ProductVariantStocksDelete",
+    "ProductVariantStocksUpdate",
+]
